Call save concert function in express server timer tick before deleting previous concert data. With this, in theory, the app works with the two servers being completely separate processes, and saves the data and audio to the server properly for review.

+++ MERN/backend/Express/src/functions/saveConcert.functions.ts
@@ -12,44 +12,48 @@ import { Op } from "sequelize";
 const saveConcert = function () {
 
     // Read from json
-    const concertData: any = JSON.parse(readFileSync(TEMP_FOLDER + "data", "utf8").toString());
+    const path = TEMP_FOLDER + "data";
+    const concertData: any = fs.existsSync(path) ? fs.readFileSync(path).toString() : "";
 
-    console_log(concertData);
+    console_log("Concert data: ", concertData);
 
-    // // Convert raw audio file wav. 
-    let wav = new WaveFile();
-    const mixedBuffer: Buffer = readFileSync(TEMP_FOLDER + "recording.bin");
-    const samples16 = new Int16Array(mixedBuffer.buffer, mixedBuffer.byteOffset, mixedBuffer.byteLength / Int16Array.BYTES_PER_ELEMENT);
-    wav.fromScratch(1, 32000, '16', samples16);
-    const fileName: string = Math.floor((Math.random() * 800000) + 100000).toString() + ".wav";
-    fs.writeFileSync(MUSIC_FOLDER + fileName, wav.toBuffer());
-    console_log("Audio converted.\n");
+    if (concertData) {
+        console_log("Starting save concert..");
+        // // Convert raw audio file wav. 
+        let wav = new WaveFile();
+        const mixedBuffer: Buffer = readFileSync(TEMP_FOLDER + "recording.bin");
+        const samples16 = new Int16Array(mixedBuffer.buffer, mixedBuffer.byteOffset, mixedBuffer.byteLength / Int16Array.BYTES_PER_ELEMENT);
+        wav.fromScratch(1, 32000, '16', samples16);
+        const fileName: string = Math.floor((Math.random() * 800000) + 100000).toString() + ".wav";
+        fs.writeFileSync(MUSIC_FOLDER + fileName, wav.toBuffer());
+        console_log("Audio converted.\n");
 
-    // Update performer names in DB.
-    updateNames(concertData.maestroName, concertData.performerNames, concertData.groupId);
+        // Update performer names in DB.
+        updateNames(concertData.maestroName, concertData.performerNames, concertData.groupId);
 
-    // Create recording.
-    recordings.create({
-        GroupID: concertData.groupId,
-        RecordingFileName: fileName
-    }).then((recording) => {
-        let newRecording = recording.dataValues;
+        // Create recording.
+        recordings.create({
+            GroupID: concertData.groupId,
+            RecordingFileName: fileName
+        }).then((recording) => {
+            let newRecording = recording.dataValues;
 
-        console_log("New recording: ", newRecording, "\n");
-    }).catch((e) => {
-        console_log("Error: ", e.message, "\n");
-    });
+            console_log("New recording: ", newRecording, "\n");
+        }).catch((e) => {
+            console_log("Error: ", e.message, "\n");
+        });
 
-    // Delete current schedule record.
-    schedules.destroy({
-        where: {
-            GroupID: concertData.groupId
-        }
-    }).then((schedule) => {
-        console_log("Schedule deleted: ", schedule, "\n");
-    }).catch((e) => {
-        console_log("Error: ", e.message, "\n");
-    });
+        // Delete current schedule record.
+        schedules.destroy({
+            where: {
+                GroupID: concertData.groupId
+            }
+        }).then((schedule) => {
+            console_log("Schedule deleted: ", schedule, "\n");
+        }).catch((e) => {
+            console_log("Error: ", e.message, "\n");
+        });
+    }
 }
 
 const updateNames = async function (maestro: string, names: string[], groupId: number): Promise<boolean> {
